HuggingFace/requirements.txt
fastapi
pandas
numpy
sentence-transformers
huggingface-hub
scikit-learn
faiss-cpu
torch
tqdm
python-dotenv
groq